+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes to ESLit
 
+### HEAD (January 13, 2017)
+
+- Added: pre-commit to development
+- Updated: Local dependencies
+- Removed: JSCS from development
+
 ### 3.0.0 (January 13, 2017)
 
 - Updated: Use `.html` as default extension